SH-215 Move rf api related class to rf_api dir, fix SocketClient
