+++ central_unit/src/mediator/mediator.h
@@ -1,11 +1,11 @@
 #pragma once
 
-#include "rf_client.h"
 #include "async_logger.h"
 #include "socket_server.h"
 #include "service_manager/service_manager.h"
 #include "socket_client.h"
-#include "api/rf_api.h"
+#include "rf_api/rf_client.h"
+#include "rf_api/rf_api.h"
 
 #include <memory>
