+++ central_unit/src/mediator/rf_api/rf_api.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../rf_types.h"
+#include "rf_types.h"
 #include "api.h"
 #include "../../core/api/internal_api.h"
