+++ central_unit/src/mediator/rf_api/rf_client.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "rf_types.h"
-#include "session.h"
-#include "rf_driver/hc12_driver.h"
+#include "../session.h"
+#include "../rf_driver/hc12_driver.h"
 #include "async_logger.h"
 
 #include <functional>
