+++ central_unit/src/mediator/rf_api/rf_types.tpp
