+++ central_unit/src/mediator/session.cpp
@@ -1,7 +1,7 @@
 #include "session.h"
-#include "rf_client.h"
+#include "rf_api/rf_client.h"
 #include "rf_driver/hc12_driver.h"
-#include "api/rf_api.h"
+#include "rf_api/rf_api.h"
 
 #include <cmath>
 #include <utility>
